delete classes in variation-commons, map.txt is the map

+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/processors/VepAnnotationProcessor.java
@@ -84,8 +84,8 @@ public class VepAnnotationProcessor implements ItemProcessor<List<EnsemblVariant
     private String getVariantInVepInputFormat(EnsemblVariant ensemblVariant) {
         return String.join("\t",
                            ensemblVariant.getChr(),
-                           Integer.toString(ensemblVariant.getStart()),
-                           Integer.toString(ensemblVariant.getEnd()),
+                           Long.toString(ensemblVariant.getStart()),
+                           Long.toString(ensemblVariant.getEnd()),
                            ensemblVariant.getRefAlt(),
                            ensemblVariant.getStrand());
     }
