+++ src/test/java/uk/ac/ebi/eva/pipeline/model/EnsemblVariantTest.java
@@ -18,7 +18,6 @@ package uk.ac.ebi.eva.pipeline.model;
 
 import org.junit.Test;
 
-import uk.ac.ebi.eva.commons.models.data.Variant;
 import uk.ac.ebi.eva.pipeline.io.mappers.VariantVcfFactory;
 
 import java.util.List;
@@ -79,7 +78,7 @@ public class EnsemblVariantTest {
         List<Variant> result = factory.create(FILE_ID, STUDY_ID, line);
         assertEquals(1, result.size());
 
-        int start = result.get(0).getStart();
+        long start = result.get(0).getStart();
         String reference = result.get(0).getReference();
         String alternate = result.get(0).getAlternate();
 
@@ -100,7 +99,7 @@ public class EnsemblVariantTest {
         List<Variant> result = factory.create(FILE_ID, STUDY_ID, line);
         assertEquals(1, result.size());
 
-        int start = result.get(0).getStart();
+        long start = result.get(0).getStart();
         String reference = result.get(0).getReference();
         String alternate = result.get(0).getAlternate();
